Send request body in tests, if provided.
Previously, the test.request function in test.js checked for a data
argument by looking at the value of `arguments.length`, but since it was
inside a closure with no arguments, the test would never return true,
even if a data argument was passed to the enclosing function.

Now, the test is whether the data argument of the enclosing function is
truthy and has a length greater than zero.