Allow POSTing of a single event.
When POSTing event data to the `/1.0/event/put` endpoint, the POSTed
JSON data can either be an array of event objects, or a single event
object. Any other kind of JSON is ignored, and a 400 Malformed Request
response is sent to the client.

Test for a 400 response when POSTing either bad JSON or JSON whose root
value is not an array or object, and a 200 response otherwise.