Display improvements.
Fix #5, allowing the y-axis to negative. If only nonpositive values are seen,
then the y-maximum is 0; if only nonnegative values are seen, then the y-minimum
is 0; otherwise, the y-domain is the minimum and maximum value. X-labels are
always drawn on the bottom.

This also adopts D3 2.4's si-prefix format for both y-axis labels and for
rolling sum values. The y-axis labels use d3.formatPrefix to determine the SI
prefix that would be used for the largest value, and then applies that to all
ticks for consistent labeling.