Allow HTTP GET for metrics.
You can now issue a simple HTTP GET to retrieve metrics from Cube synchronously,
rather than fetching metrics asynchronously via WebSockets. The parameters are
identical to the WebSockets API. So,

    {
      "expression": "sum(random)",
      "start": "2012-01-09T01:23:00Z",
      "stop": "2012-01-10T04:56:00Z",
      "step": 300000
    }

Is equivalent to:

   http://localhost:1081/1.0/metric
     ?expression=sum(random)
     &start=2012-01-09T01:23:00Z
     &stop=2012-01-10T04:56:00Z
     &step=300000

The results are returned as a JSON array in chronological order. If an error
occurs, an empty array of results are returned. Fixes #24.