Use _id for metrics collections.
The metrics collections are capped, so an _id is not required. However, starting
in MongoDB 2.0, you can no longer perform a multi-update without an object _id,
which caused metrics invalidation to fail. The fix is relatively simple, since
the metrics values already had an implicit unique id (expression, level, time
and optional group key): make the id explicit.

Note that this change is not backwards-compatible!