Replace 20-second tier with 10-second tier.
This matches Graphite's default, which is useful for comparing metrics from
multiple backends. This also disables computing 5-minute metrics from the 10-
second tier, since for low-frequency events this is vastly more expensive than
computing the 5-minute metrics from the events directly. 10-second metrics are
still available if desired but they are no longer created as a side-effect of
requesting 5-minute metrics. Ideally, Cube could determine automatically whether
to descend to the next tier based on the number of events in each time interval,
but that'll require a bit more effort to implement.