Serialize metric computation.
Now when multiple requests for the same metrics are received concurrently, only
the first request will do actual work, and the other requests will return the
cached results.

This commit required two non-backwards-compatible changes. It removes group
query support (which was only partially supported, and could crash the server by
running out of memory), and the metrics WebSocket endpoint now emits `null` when
the last metric is reported. This makes it easier for clients to detect when all
metrics are available.

In the future, I'd like to reconsider grouped queries, perhaps by providing a
new endpoint that returns metadata about events. For example, it could query
events and return the cardinality of different properties, or the top values.
Given the metadata, the client could then generate the desired queries and
choose the appropriate visualizations.