Detect collectd's high-resolution time format.
There was a brief interval during which collectd used the high-resolution time
format in conjunction with JSON. That is no longer the case. To support both
versions, we must now guess which format collectd is using. If the specified
date is more than ~31.7 years in the future (1e12 milliseconds), then the date
is assumed to be high-resolution and divided by 1073741824. Otherwise, the date
is assumed to be in seconds since UNIX epoch.