Fix #31 - Create collections automatically.
By choosing some reasonable defaults, we can allow Cube to create collections
for both events and metrics automatically, with the right indexes. You can still
create collections by hand if you want to override these defaults (for example,
adding custom indexes to the event collections, or increasing the size of the
capped metrics collections). But this is really nice for working out of the box;
it just works!