Use Cube for logging server requests.
Rather than use util.log, we now record server requests as cube_request events,
similar to the cube_compute events added in @8177510. This records the time,
method, path and remote IP address of each request, and allows you to query Cube
to see usage. (For example, you can use the /event endpoint to watch Cube
requests in realtime, or use the /metric endpoint to build a dashboard that
monitors Cube.) I elected not to record every message sent to the WebSocket
endpoints, as this would duplicate every event PUT to /event.