Allow event stream delay to be overridden.
By default there is a five-second delay when event streaming, to give events
time to arrive before they are skipped by listeners. You can now change the
stream delay as part of the /event/get request. This is useful if there is a lag
for incoming events (say because of polling a primary data source), or if you
want a time-shifted stream of old events (say, those from yesterday).