Replace Cube front-end with node-static.
The Cube front-end is now just a static file server (in addition to the /event
and /metric endpoints that power the collector and evaluator). This removes the
visualization component from Cube, so that it can focus on data collection and
metric computation. To make visualizations, you now simply write a bit of HTML
and JavaScript to fetch and display metrics. This commit will followup with some
examples that demonstrate the technique.

Also, add a 1-minute tier, and remove the week and year tiers. The highest tier
is now the 1-day tier. We could still use the old tiers, but there's a little
bit of overhead to invalidate every tier for every event; perhaps in the future,
we'll want to configure tiers on a per-event basis.