Fix a bug with metric invalidation.
This reverts @76a7527. That change introduced a bug because I forgot {multi:
true}, so the collector was only invalidating the first matching first metric
for a given event, rather than all of them.

Also, by delaying metric invalidation, the evaluator is less likely to encounter
the race condition where a new event is collected between the evaluator reading
the events and saving the computed metric. Still, this solution is not perfect.